+++ packages/schema/src/Schema.ts
@@ -99,7 +99,7 @@ export const make = <A, I = A, R = never>(ast: AST.AST): SchemaClass<A, I, R> =>
     static [TypeId] = variance
     static ast = ast
     static annotations(annotations: Annotations.Schema<A>) {
-      return make<A, I, R>(AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return make<A, I, R>(mergeSchemaAnnotations(this.ast, annotations))
     }
     static pipe() {
       return pipeArguments(this, arguments)
@@ -118,8 +118,12 @@ const variance = {
   _R: (_: never) => _
 }
 
-const toASTAnnotations = (
-  annotations?: Record<string | symbol, any> | undefined
+interface AllAnnotations<A, TypeParameters extends ReadonlyArray<any>>
+  extends Annotations.Schema<A, TypeParameters>, PropertySignature.Annotations<A>
+{}
+
+const toASTAnnotations = <A, TypeParameters extends ReadonlyArray<any>>(
+  annotations?: AllAnnotations<A, TypeParameters>
 ): AST.Annotations => {
   if (!annotations) {
     return {}
@@ -165,6 +169,9 @@ const toASTAnnotations = (
 
   return out
 }
+
+const mergeSchemaAnnotations = <A>(ast: AST.AST, annotations: Annotations.Schema<A>): AST.AST =>
+  AST.annotations(ast, toASTAnnotations(annotations))
 
 /**
  * @category annotations
@@ -323,6 +330,8 @@ export const typeSchema = <A, I, R>(schema: Schema<A, I, R>): SchemaClass<A> => 
 /* c8 ignore start */
 export {
   /**
+   * By default the option `exact` is set to `true`.
+   *
    * @throws `ParseError`
    * @category validation
    * @since 0.67.0
@@ -373,6 +382,8 @@ export {
    */
   encodeUnknownSync,
   /**
+   * By default the option `isExact` is set to `true`.
+   *
    * @category validation
    * @since 0.67.0
    */
@@ -580,7 +591,7 @@ export interface Literal<Literals extends array_.NonEmptyReadonlyArray<AST.Liter
 
 const getDefaultLiteralAST = <Literals extends array_.NonEmptyReadonlyArray<AST.LiteralValue>>(
   literals: Literals
-): AST.AST =>
+) =>
   AST.isMembers(literals)
     ? AST.Union.make(AST.mapMembers(literals, (literal) => new AST.Literal(literal)))
     : new AST.Literal(literals[0])
@@ -591,7 +602,7 @@ const makeLiteralClass = <Literals extends array_.NonEmptyReadonlyArray<AST.Lite
 ): Literal<Literals> =>
   class LiteralClass extends make<Literals[number]>(ast) {
     static override annotations(annotations: Annotations.Schema<Literals[number]>): Literal<Literals> {
-      return makeLiteralClass(this.literals, AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeLiteralClass(this.literals, mergeSchemaAnnotations(this.ast, annotations))
     }
     static literals = [...literals] as Literals
   }
@@ -652,7 +663,7 @@ export interface Enums<A extends EnumsDefinition> extends AnnotableClass<Enums<A
  */
 export type EnumsDefinition = { [x: string]: string | number }
 
-const getDefaultEnumsAST = <A extends EnumsDefinition>(enums: A): AST.AST =>
+const getDefaultEnumsAST = <A extends EnumsDefinition>(enums: A) =>
   new AST.Enums(
     Object.keys(enums).filter(
       (key) => typeof enums[enums[key]] !== "number"
@@ -665,7 +676,7 @@ const makeEnumsClass = <A extends EnumsDefinition>(
 ): Enums<A> =>
   class EnumsClass extends make<A[keyof A]>(ast) {
     static override annotations(annotations: Annotations.Schema<A[keyof A]>) {
-      return makeEnumsClass(this.enums, AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeEnumsClass(this.enums, mergeSchemaAnnotations(this.ast, annotations))
     }
 
     static enums = { ...enums }
@@ -1039,7 +1050,7 @@ export interface Union<Members extends ReadonlyArray<Schema.Any>> extends
   annotations(annotations: Annotations.Schema<Schema.Type<Members[number]>>): Union<Members>
 }
 
-const getDefaultUnionAST = <Members extends ReadonlyArray<Schema.Any>>(members: Members): AST.AST =>
+const getDefaultUnionAST = <Members extends ReadonlyArray<Schema.Any>>(members: Members) =>
   AST.Union.members(members.map((m) => m.ast))
 
 const makeUnionClass = <Members extends ReadonlyArray<Schema.Any>>(
@@ -1050,7 +1061,7 @@ const makeUnionClass = <Members extends ReadonlyArray<Schema.Any>>(
     extends make<Schema.Type<Members[number]>, Schema.Encoded<Members[number]>, Schema.Context<Members[number]>>(ast)
   {
     static override annotations(annotations: Annotations.Schema<Schema.Type<Members[number]>>): Union<Members> {
-      return makeUnionClass(this.members, AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeUnionClass(this.members, mergeSchemaAnnotations(this.ast, annotations))
     }
 
     static members = [...members] as any as Members
@@ -1265,7 +1276,7 @@ export interface TupleType<Elements extends TupleType.Elements, Rest extends Tup
 const getDefaultTupleTypeAST = <Elements extends TupleType.Elements, Rest extends TupleType.Rest>(
   elements: Elements,
   rest: Rest
-): AST.AST =>
+) =>
   new AST.TupleType(
     elements.map((el) => isSchema(el) ? new AST.OptionalType(el.ast, false) : el.ast),
     rest.map((el) => isSchema(el) ? new AST.Type(el.ast) : el.ast),
@@ -1285,7 +1296,7 @@ const makeTupleTypeClass = <Elements extends TupleType.Elements, Rest extends Tu
     static override annotations(
       annotations: Annotations.Schema<TupleType.Type<Elements, Rest>>
     ): TupleType<Elements, Rest> {
-      return makeTupleTypeClass(this.elements, this.rest, AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeTupleTypeClass(this.elements, this.rest, mergeSchemaAnnotations(this.ast, annotations))
     }
 
     static elements = [...elements] as any as Elements
@@ -1328,7 +1339,7 @@ export interface Array$<Value extends Schema.Any> extends TupleType<[], [Value]>
 const makeArrayClass = <Value extends Schema.Any>(value: Value, ast?: AST.AST): Array$<Value> =>
   class ArrayClass extends makeTupleTypeClass<[], [Value]>([], [value], ast) {
     static override annotations(annotations: Annotations.Schema<TupleType.Type<[], [Value]>>) {
-      return makeArrayClass(this.value, AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeArrayClass(this.value, mergeSchemaAnnotations(this.ast, annotations))
     }
 
     static value = value
@@ -1356,7 +1367,7 @@ export interface NonEmptyArray<Value extends Schema.Any> extends TupleType<[Valu
 const makeNonEmptyArrayClass = <Value extends Schema.Any>(value: Value, ast?: AST.AST): NonEmptyArray<Value> =>
   class NonEmptyArrayClass extends makeTupleTypeClass<[Value], [Value]>([value], [value], ast) {
     static override annotations(annotations: Annotations.Schema<TupleType.Type<[Value], [Value]>>) {
-      return makeNonEmptyArrayClass(this.value, AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeNonEmptyArrayClass(this.value, mergeSchemaAnnotations(this.ast, annotations))
     }
 
     static value = value
@@ -1512,7 +1523,7 @@ export class PropertySignatureTransformation {
   }
 }
 
-const propertySignatureAnnotations_ = (
+const mergeSignatureAnnotations = (
   ast: PropertySignature.AST,
   annotations: AST.Annotations
 ): PropertySignature.AST => {
@@ -1609,7 +1620,7 @@ class PropertySignatureImpl<
   annotations(
     annotations: PropertySignature.Annotations<Type>
   ): PropertySignature<TypeToken, Type, Key, EncodedToken, Encoded, HasDefault, R> {
-    return new PropertySignatureImpl(propertySignatureAnnotations_(this.ast, toASTAnnotations(annotations)))
+    return new PropertySignatureImpl(mergeSignatureAnnotations(this.ast, toASTAnnotations(annotations)))
   }
 
   toString() {
@@ -1649,7 +1660,7 @@ class PropertySignatureWithFromImpl<
     annotations: PropertySignature.Annotations<Type>
   ): PropertySignatureWithFromImpl<From, TypeToken, Type, Key, EncodedToken, Encoded, HasDefault, R> {
     return new PropertySignatureWithFromImpl(
-      propertySignatureAnnotations_(this.ast, toASTAnnotations(annotations)),
+      mergeSignatureAnnotations(this.ast, toASTAnnotations(annotations)),
       this.from
     )
   }
@@ -2224,10 +2235,13 @@ export const optional: {
     options?: Options
   ): (self: S) => [undefined] extends [Options] ? optional<S> : optionalWithOptions<S, Options>
   <S extends Schema.All, Options extends OptionalOptions<Schema.Type<S>>>(
-    self: S,
+    self: Schema.All extends S ? "you can't apply optional implicitly, use optional() instead" : S,
     options?: Options
   ): [undefined] extends [Options] ? optional<S> : optionalWithOptions<S, Options>
 } = dual((args) => isSchema(args[0]), (from, options) => {
+  // Note: `Schema.All extends S ? "you can't...` is used to prevent the case where `optional` is implicitly applied.
+  // For example: `S.String.pipe(S.optional)` would result in `S.String` being inferred as `Schema.All`,
+  // which is not the intended behavior. This is mostly an aesthetic consideration, so if it causes issues, we can remove it.
   return new PropertySignatureWithFromImpl(optionalPropertySignatureAST(from, options), from)
 })
 
@@ -2419,7 +2433,7 @@ const isPropertySignature = (u: unknown): u is PropertySignature.All =>
 const getDefaultTypeLiteralAST = <
   Fields extends Struct.Fields,
   const Records extends IndexSignature.Records
->(fields: Fields, records: Records): AST.AST => {
+>(fields: Fields, records: Records) => {
   const ownKeys = util_.ownKeys(fields)
   const pss: Array<AST.PropertySignature> = []
   if (ownKeys.length > 0) {
@@ -2528,7 +2542,7 @@ const makeTypeLiteralClass = <
     static override annotations(
       annotations: Annotations.Schema<Types.Simplify<TypeLiteral.Type<Fields, Records>>>
     ): TypeLiteral<Fields, Records> {
-      return makeTypeLiteralClass(this.fields, this.records, AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeTypeLiteralClass(this.fields, this.records, mergeSchemaAnnotations(this.ast, annotations))
     }
 
     static fields = { ...fields }
@@ -2648,7 +2662,7 @@ const makeRecordClass = <K extends Schema.All, V extends Schema.All>(key: K, val
     static override annotations(
       annotations: Annotations.Schema<Types.Simplify<TypeLiteral.Type<{}, [{ key: K; value: V }]>>>
     ) {
-      return makeRecordClass(key, value, AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeRecordClass(key, value, mergeSchemaAnnotations(this.ast, annotations))
     }
 
     static key = key
@@ -2670,8 +2684,7 @@ export const Record = <K extends Schema.All, V extends Schema.All>(key: K, value
 export const pick = <A, I, Keys extends ReadonlyArray<keyof A & keyof I>>(...keys: Keys) =>
 <R>(
   self: Schema<A, I, R>
-): SchemaClass<Types.Simplify<Pick<A, Keys[number]>>, Types.Simplify<Pick<I, Keys[number]>>, R> =>
-  make(AST.pick(self.ast, keys))
+): SchemaClass<Pick<A, Keys[number]>, Pick<I, Keys[number]>, R> => make(AST.pick(self.ast, keys))
 
 /**
  * @category struct transformations
@@ -2680,8 +2693,7 @@ export const pick = <A, I, Keys extends ReadonlyArray<keyof A & keyof I>>(...key
 export const omit = <A, I, Keys extends ReadonlyArray<keyof A & keyof I>>(...keys: Keys) =>
 <R>(
   self: Schema<A, I, R>
-): SchemaClass<Types.Simplify<Omit<A, Keys[number]>>, Types.Simplify<Omit<I, Keys[number]>>, R> =>
-  make(AST.omit(self.ast, keys))
+): SchemaClass<Omit<A, Keys[number]>, Omit<I, Keys[number]>, R> => make(AST.omit(self.ast, keys))
 
 /**
  * Given a schema `Schema<A, I, R>` and a key `key: K`, this function extracts a specific field from the `A` type,
@@ -2724,7 +2736,7 @@ export const pluck: {
   <A, I, R, K extends keyof A & keyof I>(
     schema: Schema<A, I, R>,
     key: K
-  ): Schema<A[K], Types.Simplify<Pick<I, K>>, R> => {
+  ): Schema<A[K], Pick<I, K>, R> => {
     const ps = AST.getPropertyKeyIndexedAccess(AST.typeAST(schema.ast), key)
     const value = make<A[K], A[K], R>(ps.isOptional ? AST.orUndefined(ps.type) : ps.type)
     return transform(
@@ -2761,7 +2773,7 @@ export interface brand<S extends Schema.Any, B extends string | symbol>
 const makeBrandClass = <S extends Schema.Any, B extends string | symbol>(ast: AST.AST): brand<S, B> =>
   class BrandClass extends make<Schema.Type<S> & Brand<B>, Schema.Encoded<S>, Schema.Context<S>>(ast) {
     static override annotations(annotations: Annotations.Schema<Schema.Type<S> & Brand<B>>): brand<S, B> {
-      return makeBrandClass(AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeBrandClass(mergeSchemaAnnotations(this.ast, annotations))
     }
 
     static make = (a: Brand.Unbranded<Schema.Type<S> & Brand<B>>, options?: MakeOptions): Schema.Type<S> & Brand<B> => {
@@ -3009,8 +3021,8 @@ const intersectUnionMembers = (
 export interface extend<Self extends Schema.Any, That extends Schema.Any> extends
   AnnotableClass<
     extend<Self, That>,
-    Types.Simplify<Schema.Type<Self> & Schema.Type<That>>,
-    Types.Simplify<Schema.Encoded<Self> & Schema.Encoded<That>>,
+    Schema.Type<Self> & Schema.Type<That>,
+    Schema.Encoded<Self> & Schema.Encoded<That>,
     Schema.Context<Self> | Schema.Context<That>
   >
 {}
@@ -3045,19 +3057,11 @@ export interface extend<Self extends Schema.Any, That extends Schema.Any> extend
  * @since 0.67.0
  */
 export const extend: {
-  <That extends Schema.Any>(
-    that: That
-  ): <Self extends Schema.Any>(self: Self) => extend<Self, That>
-  <Self extends Schema.Any, That extends Schema.Any>(
-    self: Self,
-    that: That
-  ): extend<Self, That>
+  <That extends Schema.Any>(that: That): <Self extends Schema.Any>(self: Self) => extend<Self, That>
+  <Self extends Schema.Any, That extends Schema.Any>(self: Self, that: That): extend<Self, That>
 } = dual(
   2,
-  <Self extends Schema.Any, That extends Schema.Any>(
-    self: Self,
-    that: That
-  ) => make(extendAST(self.ast, that.ast, []))
+  <Self extends Schema.Any, That extends Schema.Any>(self: Self, that: That) => make(extendAST(self.ast, that.ast, []))
 )
 
 /**
@@ -3143,7 +3147,7 @@ const makeRefineClass = <From extends Schema.Any, A>(
 ): refine<A, From> =>
   class RefineClass extends make<A, Schema.Encoded<From>, Schema.Context<From>>(ast) {
     static override annotations(annotations: Annotations.Schema<A>): refine<A, From> {
-      return makeRefineClass(this.from, this.filter, AST.annotations(this.ast, toASTAnnotations(annotations)))
+      return makeRefineClass(this.from, this.filter, mergeSchemaAnnotations(this.ast, annotations))
     }
 
     static from = from
@@ -3291,7 +3295,7 @@ const makeTransformationClass = <From extends Schema.Any, To extends Schema.Any,
       return makeTransformationClass<From, To, R>(
         this.from,
         this.to,
-        AST.annotations(this.ast, toASTAnnotations(annotations))
+        mergeSchemaAnnotations(this.ast, annotations)
       )
     }
 
@@ -3558,14 +3562,14 @@ export const attachPropertySignature: {
     value: V,
     annotations?: Annotations.Schema<Types.Simplify<A & { readonly [k in K]: V }>>
   ): SchemaClass<Types.Simplify<A & { readonly [k in K]: V }>, I, R> => {
-    const attached = extend(
+    const ast = extend(
       typeSchema(schema),
       Struct({ [key]: Predicate.isSymbol(value) ? UniqueSymbolFromSelf(value) : Literal(value) })
     ).ast
     return make(
       new AST.Transformation(
         schema.ast,
-        annotations ? AST.annotations(attached, toASTAnnotations(annotations)) : attached,
+        annotations ? mergeSchemaAnnotations(ast, annotations) : ast,
         new AST.TypeLiteralTransformation(
           [
             new AST.PropertySignatureTransformation(