+++ packages/opentelemetry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @effect/opentelemetry
 
+## 0.34.19
+
+### Patch Changes
+
+- Updated dependencies [[`3572646`](https://github.com/Effect-TS/effect/commit/3572646d5e0804f85bc7f64633fb95722533f9dd), [`1aed347`](https://github.com/Effect-TS/effect/commit/1aed347a125ed3847ec90863424810d6759cbc85), [`df4bf4b`](https://github.com/Effect-TS/effect/commit/df4bf4b62e7b316c6647da0271fc5544a84e7ba2), [`f085f92`](https://github.com/Effect-TS/effect/commit/f085f92dfa204afb41823ffc27d437225137643d)]:
+  - effect@3.3.2
+
 ## 0.34.18
 
 ### Patch Changes